Fix the Enable, Disable and List products
The enable and disable they were not in the ProductsController, ProductService and the IProductService interface, these methods are delcared and functional. Into the Models of Product and ProductCreateDTO the attributes of quantity should be in the model, and the attributes of createAt, updateAt and isActive are created into the model, and modify the service.